requirements.txt
django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
psycopg2
django-filter
whitenoise
gunicorn
argon2-cffi
argon2-cffi-bindings
asgiref
cffi
requests
openai